chore(character): separate converter
